add auth manager

+++ src/java/businessLogic/controller/HandleVehicleCRUD.java
@@ -20,43 +20,49 @@ public class HandleVehicleCRUD {
 
     public boolean createVehicle(String brand, Integer model, String color,
             Float cost, Float sellP, String description, String type) {
-        VehicleDAO vehicleDAO = new VehicleDAO();
-        Vehicle vehicle = new Vehicle();
-        vehicle.setBrand(brand);
-        vehicle.setModel(model);
-        vehicle.setColor(color);
-        vehicle.setCost(cost);
-        vehicle.setSellPrice(sellP);
-        vehicle.setDescription(description);
-        vehicle.setType(type);
-        vehicleDAO.persist(vehicle);
-        return true;
+        if (AuthentificationManager.checkPermissions(Vehicle.class, Actions.CREATE)) {
+            VehicleDAO vehicleDAO = new VehicleDAO();
+            Vehicle vehicle = new Vehicle();
+            vehicle.setBrand(brand);
+            vehicle.setModel(model);
+            vehicle.setColor(color);
+            vehicle.setCost(cost);
+            vehicle.setSellPrice(sellP);
+            vehicle.setDescription(description);
+            vehicle.setType(type);
+            vehicleDAO.persist(vehicle);
+            return true;
+        }
+        return false;
     }
 
     public void editVehicle(Vehicle v) {
-        try {
-            VehicleDAO vehicleDAO = new VehicleDAO();
-            vehicleDAO.edit(v);
-        } catch (NonexistentEntityException ex) {
-            Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (RollbackFailureException ex) {
-            Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (Exception ex) {
-            Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
+        if (AuthentificationManager.checkPermissions(Vehicle.class, Actions.UPDATE)) {
+            try {
+                VehicleDAO vehicleDAO = new VehicleDAO();
+                vehicleDAO.edit(v);
+            } catch (NonexistentEntityException ex) {
+                Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
+            } catch (RollbackFailureException ex) {
+                Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
+            } catch (Exception ex) {
+                Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
+            }
         }
-
     }
 
     public void deleteVehicle(Vehicle v) {
-        VehicleDAO vehicleDAO = new VehicleDAO();
-        try {
-            vehicleDAO.destroy(v.getVehicleId());
-        } catch (NonexistentEntityException ex) {
-            Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (RollbackFailureException ex) {
-            Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (Exception ex) {
-            Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
+        if (AuthentificationManager.checkPermissions(Vehicle.class, Actions.DELETE)) {
+            VehicleDAO vehicleDAO = new VehicleDAO();
+            try {
+                vehicleDAO.destroy(v.getVehicleId());
+            } catch (NonexistentEntityException ex) {
+                Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
+            } catch (RollbackFailureException ex) {
+                Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
+            } catch (Exception ex) {
+                Logger.getLogger(HandleVehicleCRUD.class.getName()).log(Level.SEVERE, null, ex);
+            }
         }
     }
 }
